arreglada codificacion UTF

+++ src/main/java/tfg/travel_with_me_a_p_i/domain/Resena.java
@@ -36,7 +36,7 @@ public class Resena {
     private String contenido;
 
     @Column
-    private Integer nota;
+    private Float nota;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "hotel_id")
