+++ src/main/java/tfg/travel_with_me_a_p_i/model/ResenaDTO.java
@@ -26,7 +26,7 @@ public class ResenaDTO {
     @Size(max = 2000)
     private String contenido;
 
-    private Integer nota;
+    private Float nota;
 
     private Long hotel;
